worked on Decimal fields

+++ src/app/rocket/si/content/impl/SiFields.php
@@ -67,7 +67,7 @@ class SiFields {
 	 * @param int|null $value
 	 * @return \rocket\si\content\impl\NumberInSiField
 	 */
-	static function numberIn(?int $value) {
+	static function numberIn(?float $value) {
 		return new NumberInSiField($value);
 	}
